+++ lib/simply_stored/couch/has_many.rb
@@ -9,18 +9,23 @@ module SimplyStored
         define_method(name) do |*args|
           options = args.first && args.first.is_a?(Hash) && args.first
           if options
-            options.assert_valid_keys(:force_reload, :with_deleted)
-            forced_reload = options[:force_reload]
+            options.assert_valid_keys(:force_reload, :with_deleted, :limit)
+            forced_reload = options.delete(:force_reload)
             with_deleted = options[:with_deleted]
+            limit = options[:limit]
           else
             forced_reload = false
             with_deleted = false
+            limit = nil
           end
 
-          if forced_reload || instance_variable_get("@#{name}").nil?
-            instance_variable_set("@#{name}", find_associated(name, self.class, :with_deleted => with_deleted))
+          cached_results = cached_results = send("_get_cached_#{name}")
+          cache_key = _cache_key_for(options)
+          if forced_reload || cached_results[cache_key].nil? 
+            cached_results[cache_key] = find_associated(name, self.class, :with_deleted => with_deleted, :limit => limit)
+            instance_variable_set("@#{name}", cached_results)
           end
-          instance_variable_get("@#{name}")
+          cached_results[cache_key]
         end
       end
       
@@ -30,26 +35,31 @@ module SimplyStored
         define_method(name) do |*args|
           options = args.first && args.first.is_a?(Hash) && args.first
           if options
-            options.assert_valid_keys(:force_reload, :with_deleted)
+            options.assert_valid_keys(:force_reload, :with_deleted, :limit)
             forced_reload = options[:force_reload]
             with_deleted = options[:with_deleted]
+            limit = options[:limit]
           else
             forced_reload = false
             with_deleted = false
+            limit = nil
           end
           
-          if forced_reload || instance_variable_get("@#{name}").nil?
+          cached_results = send("_get_cached_#{name}")
+          cache_key = _cache_key_for(options)
+          
+          if forced_reload || cached_results[cache_key].nil?
             
             # there is probably a faster way to query this
-            intermediate_objects = find_associated(through, self.class, :with_deleted => with_deleted)
+            intermediate_objects = find_associated(through, self.class, :with_deleted => with_deleted, :limit => limit)
             
             through_objects = intermediate_objects.map do |intermediate_object|
               intermediate_object.send(name.to_s.singularize.underscore, :with_deleted => with_deleted)
             end.flatten.uniq
-            
-            instance_variable_set("@#{name}", through_objects)
+            cached_results[cache_key] = through_objects
+            instance_variable_set("@#{name}", cached_results)
           end
-          instance_variable_get("@#{name}")
+          cached_results[cache_key]
         end
       end
       
@@ -59,9 +69,11 @@ module SimplyStored
           raise ArgumentError, "expected #{klass} got #{value.class}" unless value.is_a?(klass)
           
           value.send("#{self.class.foreign_key}=", id)
-          value.save
-          cached_version = instance_variable_get("@#{name}") || []
-          instance_variable_set("@#{name}", cached_version << value)
+          value.save(false)
+          
+          cached_results = send("_get_cached_#{name}")[:all]
+          send("_set_cached_#{name}", (cached_results || []) << value, :all)
+          nil
         end
       end
 
@@ -78,8 +90,9 @@ module SimplyStored
             value.save
           end
           
-          cached_version = instance_variable_get("@#{name}") || []
-          instance_variable_set("@#{name}", cached_version.delete_if{|item| item.id == value.id})
+          cached_results = send("_get_cached_#{name}")[:all]
+          send("_set_cached_#{name}", (cached_results || []).delete_if{|item| item.id == value.id}, :all)
+          nil
         end
       end
       
@@ -113,6 +126,22 @@ module SimplyStored
         end
       end
       
+      def define_cache_accessors(name)
+        define_method "_get_cached_#{name}" do
+          instance_variable_get("@#{name}") || {}
+        end
+        
+        define_method "_set_cached_#{name}" do |value, cache_key|
+          cached = send("_get_cached_#{name}")
+          cached[cache_key] = value
+          instance_variable_set("@#{name}", cached)
+        end
+        
+        define_method "_cache_key_for" do |options|
+          options.blank? ? :all : options.to_s
+        end
+      end
+      
       class Property
         attr_reader :name, :options
         
@@ -127,11 +156,13 @@ module SimplyStored
           
           if options[:through]
             owner_clazz.class_eval do
+              define_cache_accessors(name)
               define_has_many_through_getter(name, options[:through])
               define_has_many_count(name, options[:through])
             end
           else
             owner_clazz.class_eval do
+              define_cache_accessors(name)
               define_has_many_getter(name)
               define_has_many_setter_add(name)
               define_has_many_setter_remove(name)